% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_genes}
\alias{cc_genes}
\title{Cell cycle genes}
\format{
A character vector with 97 entries
}
\source{
\url{https://satijalab.org/seurat/articles/cell_cycle_vignette.html}
}
\usage{
cc_genes
}
\description{
A character vector containing cell-cycle phase-specific gene symbols; 1:43
are S-Phase specific, 44:97 are G2/M specific. Used for cell cycle scoring by
\code{\link{process_data}}
}
\concept{data}
\keyword{datasets}
