% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/il22_response.R
\name{il22_response}
\alias{il22_response}
\title{IL22-response analyses}
\usage{
il22_response(data = NULL, output_dir = "./")
}
\arguments{
\item{data}{The final Seurat object produced by running the
\code{\link{process_data}} function}

\item{output_dir}{Path to directory where output files should be saved
(default is "./")}
}
\value{
This function returns the input Seurat object with the IL22 response
score added to the metadata \code{col.name = "IL22_Module"}. Additionally,
2 output files are saved to the specified output directory: \itemize{\item
"DE_IL22.txt" is a table containing the IL22+ vs IL22- differential
expression results (Supplemental Table 3) \item "DE_IL22_perCluster.txt" is a table containing the
IL22+ vs IL22- differential expression results \strong{PER CLUSTER}}
}
\description{
Reproduces the analyess related to cellular response to IL22
}
\details{
This function first runs differential expression analysis between
IL22+ and IL22- cells by calling the \code{\link[Seurat]{FindMarkers}}
function (Wilcoxon Rank Sum test), and saves the output (Supplemental Table 3). The top 50 genes
(by Avg Log2FC) up-regulated in IL22+ cells are then used to calculate an
IL22-response module score. Finally, the differential gene expression
analysis between IL22+ and IL22- cells is repeated independently for each
cluster and the output is saved.
}
\examples{
\dontrun{
process_data() \%>\% il22_response()}
}
\concept{analysis}
