% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{LengthCheck}
\alias{LengthCheck}
\title{Length check
Internal length check used by \code{\link{testModuleScore}}}
\usage{
LengthCheck(values, cutoff = 0)
}
\description{
Length check
Internal length check used by \code{\link{testModuleScore}}
}
\author{
Tito Candelli
}
\keyword{internal}
