% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_analysis.R
\name{model_diffusion}
\alias{model_diffusion}
\title{Model trajectories using DiffusionMaps}
\usage{
model_diffusion(data = NULL)
}
\arguments{
\item{data}{Processed Seurat object: output from \code{\link{process_data}}}
}
\value{
A list of Seurat objects (per trajectory) which have DiffusionMaps
embeddings stored in a dimensional reduction slot. Accessible via
\code{SeuratObject::Embeddings(object, reduction = "dc")}
}
\description{
Subset cells per trajectory and project in DiffusionMaps space
}
\details{
This function creates a list of subsetted Seurat objects (per
trajectory) and projects the first 25 integrated principal components for each object in
DiffusionMaps space using the \code{destiny} package. The resulting cell
embeddings are then stashed in the "reductions" slot of the corresponding
Seurat objects.
}
\examples{
\dontrun{
object <- process_data()

object.list <- model_diffusion(data=object)

slingshot.models <- model_slingshot(data=object.list)}
}
\concept{analysis}
