% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_analysis.R
\name{model_slingshot}
\alias{model_slingshot}
\title{Infer trajectories using slingshot}
\usage{
model_slingshot(data = NULL)
}
\arguments{
\item{data}{Output from \code{\link{model_diffusion}}; List of Seurat objects
for which DiffusionMaps embeddings have been calculated.}
}
\value{
A list of slingshot objects corresponding to the list of Seurat
objects used as input
}
\description{
A function to run slingshot on DiffusionMaps modeled trajectories
}
\details{
This function calls slingshot on each object in the list of Seurat
objects generated by \code{\link{model_diffusion}}, using the first 2
DiffusionMaps components as input. In all cases the "ISC" cluster is used as the starting point for defining lineages
}
\examples{
\dontrun{
object <- process_data()

object.list <- model_diffusion(data=object)

slingshot.models <- model_slingshot(data=object.list)}
}
