% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanethAnalysis-package.R
\docType{package}
\name{PanethAnalysis-package}
\alias{PanethAnalysis}
\alias{PanethAnalysis-package}
\title{PanethAnalysis: Analysis of single-cell RNA-sequencing data from He, G-W. and Lin, L., et al (2022) <XX>}
\description{
This package provides the data and wrapper functions required to reproduce the analyses from our publication "Optimal human intestinal organoid model reveals interleukin-22 induced Paneth cell differentiation".
}
\author{
\strong{Maintainer}: Jeff DeMartino \email{jdemartino1@gmail.com} (\href{https://orcid.org/0000-0001-7366-4789}{ORCID})

}
\keyword{internal}
