% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_il22}
\alias{plot_il22}
\title{Plot IL22 response figures}
\usage{
plot_il22(
  object = NULL,
  dge.output = NA,
  figure = NA,
  save = FALSE,
  save.dir = "."
)
}
\arguments{
\item{object}{The output Seurat object generated by
\code{\link{il22_response}}}

\item{dge.output}{Path to the DGE file generated by
\code{\link{il22_response}} ("DE_IL22.txt")}

\item{figure}{ID of the figure to be reproduced. Accepted values are "2D",
"S7B and "S7D" (See Details for description of figure panels)}

\item{save}{Whether to save the specified plot as a pdf (default = FALSE)}

\item{save.dir}{Path to the directory where the output plot should be saved
(default = ".")}
}
\value{
The plot specified in the "figure" argument
}
\description{
Reproduces figure panels related to the IL22 response
}
\details{
This function reproduces the following figure panels: \itemize{\item
\strong{Figure 2D}: Differential gene expression (MA) plot, IL22+ vs IL22-
\item \strong{Figure S7B}: Violin plots depicting the IL22-response module
score between cells cultured in IL22+ and IL22- media (overall and per
cluster) \item \strong{Figure S7D}: Violin plots depicting normalized gene
expression of REG1A, REG1B and DMBT1 between cells cultured in IL22+ and
IL22= media (per cell-type cluster)}
}
\examples{
\dontrun{
# Make each plot and save as pdf
plot_il22(object = il22.output.obj, figure = "2D", dge.output = "DE_IL22.txt", save = T, save.dir = ".")
plot_il22(object = il22.output.obj, figure = "S7B",  save = T, save.dir = ".")
plot_il22(object = il22.output.obj, figure = "S7D",  save = T, save.dir = ".")}
}
