% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_processed}
\alias{plot_processed}
\title{Plot processed data figures}
\usage{
plot_processed(object = NULL, figure = NA, save = FALSE, save.dir = ".")
}
\arguments{
\item{object}{The output Seurat object generated by
\code{\link{process_data}}}

\item{figure}{Character ID of the figure to be reproduced. Accepted values are "2A",
"2B", "2C", "ED_3A", "ED_3B", "ED_4D", "ED_4E" and "ED_8B" (See Details for
description of figure panels)}

\item{save}{Save the specified plot as a pdf? (default = FALSE)}

\item{save.dir}{Path to the directory where the output plot should be saved
(default = ".")}
}
\value{
The plot specified in the "figure" argument
}
\description{
Reproduces processed data figure panels
}
\details{
This function reproduces the following figure panels: \itemize{\item
\strong{Figure 2A}: UMAP projection of cells cultured in IL22+ medium,
colored by cell-type cluster \item \strong{Figure 2B}: Dot plot depicting
selected marker gene expression aggregated per cell-type cluster \item
\strong{Figure 2C}: Bar plot showing the difference in cell-type cluster
proportions between cells cultured in IL22+ and IL22- media \item
\strong{Extended data Figure 3A}: UMAP plots depicting normalized gene expression of
cell-type markers in IL22+ cultured cells \item \strong{Extended data Figure 3B}: UMAP
plots depicting normalized gene expression of WNT ligands in IL22+ cultured
cells \item \strong{Extended data Figure 4D}: Integrated UMAP projection of all cells
colored by culture condition (IL22+/IL22-) \item \strong{Extended data Figure 4E}:
Alluvial plot highlighting the proportions of each enterocyte lineage cluster per
culture condition (IL22+/IL22-)
\item \strong{Extended data Figure 8B}: UMAP plots
depicting normalized gene expression of IL22 receptor genes in cells
cultured in either IL22+ or IL22- medium}
}
\examples{
\dontrun{
# Plot all figures, save all plots as pdf files in the current working directory
plot_processed(object = final.output, figure = "2A", save = T, save.dir = ".")
plot_processed(object = final.output, figure = "2B", save = T, save.dir = ".")
plot_processed(object = final.output, figure = "2C", save = T, save.dir = ".")
plot_processed(object = final.output, figure = "ED_3A", save = T, save.dir = ".")
plot_processed(object = final.output, figure = "ED_3B", save = T, save.dir = ".")
plot_processed(object = final.output, figure = "ED_4D", save = T, save.dir = ".")
plot_processed(object = final.output, figure = "ED_4E", save = T, save.dir = ".")
plot_processed(object = final.output, figure = "ED_8B", save = T, save.dir = ".")}
}
