% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_velocity.R
\name{prep_velocity}
\alias{prep_velocity}
\title{Prep data for RNA velocity analysis}
\usage{
prep_velocity(object = NULL, output.dir = ".")
}
\arguments{
\item{object}{The final Seurat object produced by running the
\code{\link{process_data}} function}

\item{output.dir}{Path to directory where output file should be saved
(default is ".")}
}
\value{
Doesn't return an object (saves output to disk)
}
\description{
Prepares spliced/unspliced count data for analysis with scvelo
}
\details{
Takes the raw spliced/unspliced count data (pre-stored as a Seurat
object, see \code{\link{velocity_raw}}) and converts it to an h5ad file
suitable for loading into python and analyzing with scvelo. Subsets to
include only "IL22+" cells, and transfers dimensional reductions (PCA and
UMAP) from the integrated, processed, Seurat object to the h5ad file.
Workflow based loosely on
\href{http://htmlpreview.github.io/?https://github.com/satijalab/seurat-wrappers/blob/master/docs/scvelo.html}{this
Seurat vignette}
}
\examples{
\dontrun{
final.output <- preprocess_data() \%>\%
integrate_data() \%>\%
process_data()
prep_velocity(object=final.output)}
}
