% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_wrappers.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess incoming raw data}
\usage{
preprocess_data(data_list = PanethAnalysis::raw_data)
}
\arguments{
\item{data_list}{Raw count data stored in
\code{\link{raw_data}}}
}
\value{
A filtered Seurat object which is ready for integration using
\code{\link{integrate_data}}
}
\description{
This function runs the first pre-processing steps on the raw count data
stored in \code{\link{raw_data}}
}
\details{

}
\examples{
\dontrun{
preprocess_data()}
}
\seealso{
Other data processing: 
\code{\link{integrate_data}()},
\code{\link{process_data}()}
}
\concept{data processing}
