% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_data}
\alias{raw_data}
\title{Raw input data}
\format{
A list containing 2 sparse matrices (of type 'dgCMatrix') where rows are genes and cols are cells
\describe{
  \item{IL22+}
  \item{IL22-}
  ...
}
}
\usage{
raw_data
}
\description{
A list containing the raw count data from CellRanger for each of the 2 culture conditons (IL22+ and IL22-)
which were read into R using the \code{Seurat::Read10X(..., strip.suffix=T)} function call
}
\concept{data}
\keyword{datasets}
