% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stressed_cells}
\alias{stressed_cells}
\title{Stressed cells}
\format{
A character vector with 37 entries
}
\usage{
stressed_cells
}
\description{
A character vector containing the IDs of cells to be excluded based
on their expression of stress-response genes
}
\concept{data}
\keyword{datasets}
