% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{velocity_raw}
\alias{velocity_raw}
\title{Raw velocity data}
\format{
A Seurat object with 4 assay slots: spliced, unspliced, ambiguous and RNA
}
\usage{
velocity_raw
}
\description{
Seurat object generated from the output loom files produced by the velocyto
command line tool. A detailed overview of how this file was generated can be
found in the package vignette.
}
\concept{data}
\keyword{datasets}
